+++ src/Request/Authorization/PayPal.php
@@ -7,17 +7,16 @@ use ArvPayoneApi\Request\ClearingTypes;
 use ArvPayoneApi\Request\Parts\Config;
 use ArvPayoneApi\Request\Parts\Customer;
 use ArvPayoneApi\Request\Parts\RedirectUrls;
-use ArvPayoneApi\Request\Parts\ShippingAddress;
 use ArvPayoneApi\Request\Parts\SystemInfo;
 use ArvPayoneApi\Request\RequestDataContract;
 use ArvPayoneApi\Request\Types;
 
 /**
- * Class Paydirect
+ * Class PayPal
  */
-class Paydirect extends AuthorizationAbstract implements RequestDataContract
+class PayPal extends AuthorizationAbstract implements RequestDataContract
 {
-    const WALLET_TYPE = 'PDT';
+    const WALLET_TYPE = 'PPE';
     /**
      * @var string
      */
@@ -26,22 +25,17 @@ class Paydirect extends AuthorizationAbstract implements RequestDataContract
      * @var RedirectUrls
      */
     private $urls;
-    /**
-     * @var ShippingAddress
-     */
-    private $shippingAddress;
 
     /**
-     * Paydirect constructor.
+     * PayPal constructor.
      *
      * @param Config $config
-     * @param string $orderId
+     * @param $orderId
      * @param int $amount
-     * @param string $currency
+     * @param $currency
      * @param Customer $customer
      * @param SystemInfo $info
      * @param RedirectUrls $urls
-     * @param ShippingAddress $shippingAddress
      */
     public function __construct(
         Config $config,
@@ -50,8 +44,7 @@ class Paydirect extends AuthorizationAbstract implements RequestDataContract
         $currency,
         Customer $customer,
         SystemInfo $info,
-        RedirectUrls $urls,
-        ShippingAddress $shippingAddress
+        RedirectUrls $urls
     ) {
         parent::__construct(
             $config,
@@ -64,7 +57,6 @@ class Paydirect extends AuthorizationAbstract implements RequestDataContract
             $currency
         );
         $this->urls = $urls;
-        $this->shippingAddress = $shippingAddress;
     }
 
     /**
@@ -85,14 +77,4 @@ class Paydirect extends AuthorizationAbstract implements RequestDataContract
     {
         return $this->wallettype;
     }
-
-    /**
-     * Getter for ShippingAddress
-     * @return ShippingAddress
-     */
-    public function getShippingAddress()
-    {
-        return $this->shippingAddress;
-    }
-
 }