+++ src/Request/PaymentTypes.php
@@ -13,6 +13,8 @@ class PaymentTypes
     const PAYONE_SOFORT = 'Sofort';
     const PAYONE_CREDIT_CARD = 'CreditCard';
     const PAYONE_DIRECT_DEBIT = 'DirectDebit';
+    const PAYONE_PAY_PAL = 'PayPal';
+    const PAYONE_PAYDIRECT = 'Paydirect';
 
     /**
      * @return mixed